+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Features
 
   * Displayed variant when running dcos --version
+  * Always auto-install core and EE plugins on `cluster setup`